# src/api/internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package store

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Exercise struct {
	ID              int32              `json:"id"`
	Name            string             `json:"name"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	MuscleGroup     string             `json:"muscle_group"`
	Equipment       string             `json:"equipment"`
	IsAccessory     bool               `json:"is_accessory"`
	CreatedByUserID *int32             `json:"created_by_user_id"`
}

type Program struct {
	ID              int32              `json:"id"`
	Name            string             `json:"name"`
	Description     string             `json:"description"`
	ProgressionKind string             `json:"progression_kind"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
}

type ProgramDay struct {
	ID        int32  `json:"id"`
	ProgramID int32  `json:"program_id"`
	Name      string `json:"name"`
	Position  int32  `json:"position"`
	Weekday   *int32 `json:"weekday"`
}

type ProgramDayAssistance struct {
	ID           int32              `json:"id"`
	UserID       int32              `json:"user_id"`
	ProgramDayID int32              `json:"program_day_id"`
	ExerciseID   int32              `json:"exercise_id"`
	Position     int32              `json:"position"`
	Sets         int32              `json:"sets"`
	Reps         int32              `json:"reps"`
	WeightLb     pgtype.Numeric     `json:"weight_lb"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type ProgramDayExercise struct {
	ID               int32          `json:"id"`
	ProgramDayID     int32          `json:"program_day_id"`
	ExerciseID       int32          `json:"exercise_id"`
	Position         int32          `json:"position"`
	Sets             int32          `json:"sets"`
	Reps             int32          `json:"reps"`
	StartingWeightLb pgtype.Numeric `json:"starting_weight_lb"`
}

type ReportRun struct {
	ID          int32              `json:"id"`
	UserID      int32              `json:"user_id"`
	PeriodKind  string             `json:"period_kind"`
	PeriodStart pgtype.Date        `json:"period_start"`
	Status      string             `json:"status"`
	Attempts    int32              `json:"attempts"`
	ClaimedAt   pgtype.Timestamptz `json:"claimed_at"`
	SentAt      pgtype.Timestamptz `json:"sent_at"`
	LastError   string             `json:"last_error"`
}

type Session struct {
	ID           int32              `json:"id"`
	ProgramDayID int32              `json:"program_day_id"`
	PerformedOn  pgtype.Date        `json:"performed_on"`
	Notes        string             `json:"notes"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
	FinishedAt   pgtype.Timestamptz `json:"finished_at"`
	UserID       *int32             `json:"user_id"`
}

type SessionSet struct {
	ID         int32          `json:"id"`
	SessionID  int32          `json:"session_id"`
	ExerciseID int32          `json:"exercise_id"`
	SetNumber  int32          `json:"set_number"`
	TargetReps int32          `json:"target_reps"`
	ActualReps *int32         `json:"actual_reps"`
	WeightLb   pgtype.Numeric `json:"weight_lb"`
	Completed  bool           `json:"completed"`
}

type User struct {
	ID               int32              `json:"id"`
	Username         string             `json:"username"`
	DisplayName      string             `json:"display_name"`
	AvatarColor      string             `json:"avatar_color"`
	PasswordHash     string             `json:"password_hash"`
	IsAdmin          bool               `json:"is_admin"`
	CreatedAt        pgtype.Timestamptz `json:"created_at"`
	UpdatedAt        pgtype.Timestamptz `json:"updated_at"`
	CurrentProgramID *int32             `json:"current_program_id"`
}

type UserAvatar struct {
	UserID    int32              `json:"user_id"`
	Mime      string             `json:"mime"`
	Bytes     []byte             `json:"bytes"`
	Etag      string             `json:"etag"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type UserSession struct {
	TokenHash  []byte             `json:"token_hash"`
	UserID     int32              `json:"user_id"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	LastSeen   pgtype.Timestamptz `json:"last_seen"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	Persistent bool               `json:"persistent"`
}