# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Iron Temple API
  version: 0.1.0
  description: >
    REST API for the Iron Temple weight-lifting tracker. All weights are in
    pounds (lb).

    Authentication is a session cookie (`it_session`), issued by
    `POST /auth/login` and revoked by `POST /auth/logout`. Every endpoint
    requires it except `/health`, the `/auth/*` endpoints, and avatar images.
    Registration is open only while no account exists: the first sign-up claims
    the install and closes the door behind it.

    Programs, days and exercises are shared — the prescription is the same for
    everyone. Sessions are per-user: what you lifted is yours, and progression
    is computed from your own history alone.

    Programs are made of days (e.g. Workout A / B); each day prescribes exercises
    with sets, reps, and a starting weight. A session is a performed instance of
    a day, with one logged set per prescribed set.

    Progression is linear and computed server-side from session history:
    +5 lb per session on squat, bench, row, and overhead press; +10 lb on
    deadlift; a 10% deload after three consecutive failed sessions on a lift.
servers:
  - url: /api/v1

# Applies to every operation; the handful of public ones override with
# `security: []`.
security:
  - cookieAuth: []
...
paths:
...
  /exercises:
    get:
      tags: [exercises]
      operationId: listExercises
      summary: List the exercise library
      description: >
        The seeded catalogue plus the caller's own custom movements. Someone
        else's custom exercise is never returned.
      parameters:
        - name: scope
          in: query
          required: false
          schema:
            type: string
            enum: [all, performed]
            default: all
          description: >
            all returns the whole library. performed narrows it to lifts the
            caller has actually logged, which is what the Progress page wants —
            a library's worth of untrained movements reading "no sessions yet"
            is noise.
      responses:
        "200":
          description: Exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [exercises]
      operationId: createExercise
      summary: Add a custom exercise to the library
      description: >
        Creates a movement visible only to the caller. Names are compared
        case-insensitively against the shared catalogue and the caller's own
        exercises; a collision with either is a 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exercise"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises/{exerciseId}:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    delete:
      tags: [exercises]
      operationId: deleteExercise
      summary: Delete one of your own custom exercises
      description: >
        Seeded exercises cannot be deleted by anyone. A custom one cannot be
        deleted while it has logged sets against it — that would take a lift out
        of a finished session, which is a record — or while it is on a program
        day as assistance; remove it there first.
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /exercises/{exerciseId}/history:
    parameters:
      - name: exerciseId
        in: path
        required: true
        schema:
          type: integer
          format: int64
    get:
      tags: [exercises]
      operationId: getExerciseHistory
      summary: Per-session weight history for a lift (oldest first)
      responses:
        "200":
          description: History points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseHistoryPoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: it_session
      description: >
        Opaque session token. Set by the server as an HttpOnly, SameSite=Lax
        cookie — it is never readable by, or set from, JavaScript, so clients
        neither send nor manage it explicitly.
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not signed in, or the session has expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request is valid but collides with existing state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: not_found
        message:
          type: string
          example: session 42 not found
    MuscleGroup:
      type: string
      enum: [chest, back, legs, shoulders, arms, core, other]
    Equipment:
      type: string
      enum: [barbell, dumbbell, machine, cable, bodyweight, other]
    Exercise:
      description: One movement in the library.
      type: object
      required: [id, name, muscleGroup, equipment, isAccessory, isCustom]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Squat
        muscleGroup:
          $ref: "#/components/schemas/MuscleGroup"
        equipment:
          $ref: "#/components/schemas/Equipment"
        isAccessory:
          type: boolean
          description: >-
            False for the compound lifts the seeded programs prescribe, true for
            accessory work. The library leans on this to lead with the movements
            a lifter is shopping for when adding assistance.
        isCustom:
          type: boolean
          description: >-
            True for a movement this lifter created. Only custom exercises can be
            deleted, and only by their owner; the seeded catalogue is shared.
    CreateExerciseRequest:
      type: object
      required: [name, muscleGroup, equipment]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
          example: Copenhagen Plank
        muscleGroup:
          $ref: "#/components/schemas/MuscleGroup"
        equipment:
          $ref: "#/components/schemas/Equipment"
    ExerciseHistoryPoint:
      description: One performed session's top set for a lift.
      type: object
      required: [performedOn, weightLb, reps, completed]
      properties:
        performedOn:
          type: string
          format: date
        weightLb:
          type: number
          minimum: 0
        reps:
          type: integer
        completed:
          type: boolean
...